Move lib/client/BaseSubscriptionGraphqlPayload.js below lib/client/graphql/
